Final solution to The Trip - still doesn't pass
I have compared my solution to other solutions and it seems to be
working correctly, but still doesn't pass the online judging.
I have no idea why this is the case. I may come back to this in the
future and figure out why the judge doesn't like it, but for now
I am done with this problem.